revisione data creazione, gestione attivazione scheduler e recupero status da DB se presenti

+++ src/main/java/it/myhouse/switchbot/service/scheduler/DeviceStatusScheduler.java
@@ -5,6 +5,7 @@ import it.myhouse.switchbot.repository.DeviceStatusRepository;
 import it.myhouse.switchbot.service.SwitchBotApiService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
@@ -19,6 +20,9 @@ public class DeviceStatusScheduler {
     private final DeviceStatusRepository repository;
     private final SwitchBotApiService service;
 
+    @Value("${ENABLE_SCHEDULER:false}")  // di default false
+    private boolean schedulerEnabled;
+
     public DeviceStatusScheduler(SwitchBotApiService service, DeviceStatusRepository repository) {
         this.service = service;
         this.repository = repository;
@@ -30,6 +34,11 @@ public class DeviceStatusScheduler {
     @Scheduled(fixedRate = 300_000)
     public void fetchDevicesStatusPeriodically() {
         logger.info("Esecuzione schedulata: fetchDevicesStatusPeriodically");
+
+        if (!schedulerEnabled) {
+            logger.debug("Scheduler disabilitato!");
+            return; // scheduler disabilitato, esco subito
+        }
 
         try {
             List<DeviceStatus> statuses = service.getDevicesStatus();
@@ -48,13 +57,16 @@ public class DeviceStatusScheduler {
     public void cleanOldData() {
         logger.info("Esecuzione schedulata: cleanOldData");
 
+        if (!schedulerEnabled) {
+            logger.debug("Scheduler disabilitato!");
+            return; // scheduler disabilitato, esco subito
+        }
+
         try {
-            long oneWeekAgo = Instant.now()
-                    .minusSeconds(7L * 24 * 3600)
-                    .toEpochMilli();
+            Instant oneWeekAgo = Instant.now().minusSeconds(7L * 24 * 3600);
 
             int deletedRows = repository.deleteOlderThan(oneWeekAgo);
-            logger.debug("Eliminati {} record più vecchi di una settimana", deletedRows);
+            logger.debug("Eliminati {} record piu' vecchi di una settimana", deletedRows);
         } catch (Exception e) {
             logger.error("Errore durante l'eliminazione dei dati vecchi", e);
         }
